requirements.txt
sanic~=20.6.3
PyMySQL==0.9.2
aiomysql==0.0.20
cffi==1.14.3
cryptography==3.1.1
pycparser==2.20
aiocassandra==2.0.1
async-generator==1.10
cassandra-driver==3.24.0
geomet==0.2.1.post1